fix(useConversationInfo): rename to ts
